# docs/swagger.yaml
basePath: /
definitions:
  auth.LoginBody:
    properties:
      Email:
        type: string
      Password:
        type: string
    required:
    - Email
    - Password
    type: object
  auth.RefreshTokenBody:
    properties:
      RefreshToken:
        type: string
    required:
    - RefreshToken
    type: object
  auth.SignupBody:
    properties:
      Email:
        type: string
      Password:
        maxLength: 16
        minLength: 6
        type: string
    required:
    - Email
    - Password
    type: object
  models.Campaign:
    properties:
      CurrentAirdrop:
        type: integer
      Description:
        type: string
      Name:
        type: string
      Quizzes:
        items:
          type: string
        type: array
      RequiredFlaq:
        type: integer
      TickerName:
        type: string
      TotalAirdrop:
        type: integer
    required:
    - CurrentAirdrop
    - Description
    - Name
    - RequiredFlaq
    - TickerName
    - TotalAirdrop
    type: object
  models.Question:
    properties:
      AnswerIndex:
        type: integer
      Description:
        type: string
      Options:
        items:
          type: string
        type: array
      Question:
        type: string
    required:
    - AnswerIndex
    - Description
    - Options
    - Question
    type: object
  models.QuizTemplate:
    properties:
      Questions:
        items:
          $ref: '#/definitions/models.Question'
        type: array
      Title:
        type: string
      id:
        type: string
    required:
    - Questions
    - Title
    type: object
  payments.registerPaymentBody:
    properties:
      Amount:
        type: string
    required:
    - Amount
    type: object
  users.ApplyReferralBody:
    properties:
      ReferralCode:
        type: string
    required:
    - ReferralCode
    type: object
info:
  contact:
    email: ashwin@flaq.club
    name: Ashwin Prasad
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://flaq.club/privacy-policy
  title: Flaq API
  version: "2.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Enter login details
        in: body
        name: LoginBody
        required: true
        schema:
          $ref: '#/definitions/auth.LoginBody'
      produces:
      - application/json
      responses: {}
      summary: User login
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Add Data
        in: body
        name: SignupBody
        required: true
        schema:
          $ref: '#/definitions/auth.SignupBody'
      produces:
      - application/json
      responses: {}
      summary: User signup
      tags:
      - Auth
  /auth/token/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: RefreshTokenBody
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenBody'
      produces:
      - application/json
      responses: {}
      summary: Issue a new Access token
      tags:
      - Auth
  /campaign:
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign Details
        in: body
        name: models.Campaign
        required: true
        schema:
          $ref: '#/definitions/models.Campaign'
      responses: {}
      summary: Create a campaign
      tags:
      - Campaigns
  /campaign/quiz/template:
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign Details
        in: body
        name: models.QuizTemplate
        required: true
        schema:
          $ref: '#/definitions/models.QuizTemplate'
      responses: {}
      summary: Create a quiz template
      tags:
      - Campaigns
  /payments:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Register a new UPI Payment
      tags:
      - Payments
  /payments/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Register Payment
        in: body
        name: registerPaymentBody
        required: true
        schema:
          $ref: '#/definitions/payments.registerPaymentBody'
      produces:
      - application/json
      responses: {}
      summary: Register a new UPI Payment
      tags:
      - Payments
  /users/apply-referral:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add Referral Data
        in: body
        name: ApplyReferralBody
        required: true
        schema:
          $ref: '#/definitions/users.ApplyReferralBody'
      responses: {}
      summary: Apply a referral code
      tags:
      - Users
  /users/profile:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get user profile
      tags:
      - Users
schemes:
- http
swagger: "2.0"